+++ cmd/forklift/plt/cli.go
@@ -356,6 +356,8 @@ func makeModifySubcmds(versions Versions) []*cli.Command {
 			},
 		},
 		makeModifyPltSubcmds(versions),
+		// TODO: add `add-imp`, `rm-imp`, `set-imp-disabled`, `unset-imp-disabled`,
+		// `add-imp-mod`, and `rm-imp-mod` subcommands
 		makeModifyRepoSubcmds(versions),
 		makeModifyDeplSubcmds(versions),
 	)