+++ CHANGELOG.md
@@ -17,15 +17,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (cli) Added a `[dev] plt ls-imp` command to list all file import groups declared by the local/development pallet.
 - (cli) Added a `[dev] plt show-imp` command to show the specified file import group declared by the local/development pallet.
 - (cli) Now the `[dev] plt cache-all` command, and all commands which can cache staging requirements, will cache pallets required by the local/development pallet.
+- (cli) Now the `[dev] plt show-pkg` and `cache show-pkg` commands also print information about file exports.
 
 ### Changed
 
-- (cli) Now the commands for viewing a pallet/repo (e.g. `[dev] plt show`) truncate the pallet/repo's readme file to the first ten lines, to prevent long readme files from clogging up stdout.
+- (cli) Now the commands for viewing a pallet/repo (e.g. `[dev] plt show`) truncate the printout of the pallet/repo's readme file to the first ten lines of the file, to prevent long readme files from clogging up the command output.
 - (cli) Now the `[dev] plt rm-repo` command only deletes the version lock file for the specified repository, instead of deleting the entire subdirectory for the repository.
 
 ### Fixed
 
-- (spec) Fixed an incorrect example for the `target` field of the file exports object in the packaging spec.
+- (spec) Fixed an incorrect example for the `target` field of the file export object in the packaging spec.
+- (spec) Fixed a formatting error in the description of the `url` field of the file export object
+- (cli) Fixed a regression where the `[dev] plt ls-pkg` command failed with a `cache is nil` error on pallets which are not also package repositories.
 
 ## 0.8.0-alpha.0 - 2024-07-05